requirements.txt
fire==0.2.1
Pillow==6.2.1
